+++ src/components/App/App.component.jsx
@@ -3,7 +3,6 @@ import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
 import AuthProvider from '../../providers/Auth';
 import HomePage from '../../pages/Home';
-import LoginPage from '../../pages/Login';
 import NotFound from '../../pages/NotFound';
 import SecretPage from '../../pages/Secret';
 import Private from '../Private';
@@ -36,9 +35,6 @@ function App() {
           <Switch>
             <Route exact path="/">
               <HomePage />
-            </Route>
-            <Route exact path="/login">
-              <LoginPage />
             </Route>
             <Private exact path="/secret">
               <SecretPage />
